Create ess private layer
This layer is a clone of the official spacemacs ess layer, available in
layers/+lang/ess. The modifications I've made are in packages.el only,
to add Rmd support with polymode and alter ESS defaults.

Basically add polymode to the list of packages to be installed fromm
ELPA, and configure the Rmd support at the end of this file.

The other important modification is to set R coding standards in place
of Hadley's coding standards.